+++ sdk/ai/Azure.AI.Projects.OpenAI/src/Custom/OpenAI/ProjectConversationUpdateOptions.cs
@@ -1,8 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using Azure.AI.Projects.OpenAI;
-
 namespace Azure.AI.Projects.OpenAI;
 
 [CodeGenType("UpdateConversationRequest")]